minui: accept RGBA and treat it as RGBX
There are devices with RGBA_8888 pixel format.

There are also recoveries like TWRP that have
alpha support and hence for a RGBA device, I expect
to be able to set TARGET_RECOVERY_PIXEL_FORMAT = RGBA_8888

It should be supported on BoardConfig to set the
correct PixelFormat for a device.
Although recovery does not have alpha, so it can
consider RGBA devices as RGBX, but it could do so
in future.

Change-Id: Id1f3b2fbfc8ad4cf0422614580e420196614f792
Signed-off-by: Arnav Gupta <championswimmer@gmail.com>
Signed-off-by: darjwx <darjwx@gmail.com>